requirements.txt
dash>=2.17
plotly>=5.22
xarray>=2024.1
cfgrib>=0.9.12
herbie-data>=2024.1
s3fs>=2024.2
netCDF4>=1.7
h5netcdf>=1.3
h5py>=3.7
numpy>=1.26
pandas>=2.2
requests>=2.31
apscheduler>=3.10
scipy>=1.12
cartopy>=0.23
Pillow>=10.2
